Support up to 32kB of payload using 16-bit frame length
Control byte `rx_control` is only used to save the original payload
length as received in the reader, that is why it is renamed to
`rx_length`. `rx_frame_length` is assigned to the same value at the
beginning, but it is used as a payload data counter and decreased for
every byte of received payload.

Maximum size is limited to 0x7FFF instead of 0xFFFF to avoid changing
offset variables from uint16_t to uint32_t but still prevent overflows
when incrementing them.